fix getStaticProps for producthub
