+++ features/productHub/views/ProductHubRouteHandler.tsx
@@ -1,13 +1,12 @@
 import { AnimatedWrapper } from 'components/AnimatedWrapper'
 import { WithConnection } from 'components/connectWallet'
 import { AppLayout } from 'components/Layouts'
-import { getProductHubStaticProps } from 'features/productHub/helpers/getProductHubStaticProps'
 import { ALL_ASSETS, productHubOptionsMap } from 'features/productHub/meta'
 import { ProductHubProductType } from 'features/productHub/types'
 import { ProductHubView } from 'features/productHub/views'
 import { WithChildren } from 'helpers/types'
 import { useFeatureToggle } from 'helpers/useFeatureToggle'
-import { GetStaticPaths, GetStaticProps } from 'next'
+import { GetStaticPaths } from 'next'
 import React from 'react'
 
 function ProductHubRouteHandler({
@@ -54,8 +53,4 @@ export const getStaticPaths: GetStaticPaths = async ({ locales }) => {
     paths,
     fallback: false,
   }
-}
-
-export const getStaticProps: GetStaticProps = async ({ locale, params }) => {
-  return await getProductHubStaticProps(locale, params)
 }
