+++ pages/[networkOrProduct]/index.tsx
@@ -1,7 +1,7 @@
-import { getProductHubStaticProps } from 'features/productHub/helpers/getProductHubStaticProps'
 import { ProductHubProductType } from 'features/productHub/types'
 import ProductHubRouteHandler from 'features/productHub/views/ProductHubRouteHandler'
 import { GetStaticPaths, GetStaticProps } from 'next'
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
 export default ProductHubRouteHandler
 
@@ -23,5 +23,13 @@ export const getStaticPaths: GetStaticPaths = async ({ locales }) => {
 }
 
 export const getStaticProps: GetStaticProps = async ({ locale, params }) => {
-  return await getProductHubStaticProps(locale, params)
+  const product = params?.networkOrProduct as ProductHubProductType
+  const token = params?.token
+  return {
+    props: {
+      ...(await serverSideTranslations(locale || 'en', ['common'])),
+      ...(product && { product }),
+      ...(token && { token }),
+    },
+  }
 }
